fix broken min hunk size
It still a bit broken: if a chunk of 6 lines was copied 3 lines in one
file and 3 other lines in another, than cccv will still pick it up even
if min-hunk-size is, say, 4. That is unexpected.

But I am not sure how often things like that happen in practice. And I
can't think of an easy fix.